Move match_basic test to base ladder
